chore: from __future__ import annotations
This was annoyingly hard. The future import breaks dataclasses.fields().
Some test regressions resulted, and some modules now _can't_ do the
future thing. This was sloppy and maybe foolish, but 3.8 support is
slipping into the rearview so I'm okay with it for now. Spending excess
time making this super crisp is blocking release of codegen features.