feat: remove two-step-api exceptions
this is also a silly hacky excuse to yank 3.0.1. i accidentally included
breaking changes in there due to a commit message that i failed to add
the breaking changes footer to _and_ accidentally pushed directly to
master. oops! anyway.

BREAKING CHANGE: Not only will the two-step API not work anymore, but
- **Python 3.7 is no longer supported**
- code generation is a feature!
  - Render() and friends are a thing
- cases try to deserialize input as json